+++ src/main/java/uz/epam/webproject/controller/command/impl/DeleteUserCommand.java
@@ -8,6 +8,7 @@ import uz.epam.webproject.controller.command.Router;
 import uz.epam.webproject.controller.command.exception.CommandException;
 import uz.epam.webproject.entity.medicine.Medicine;
 import uz.epam.webproject.entity.user.User;
+import uz.epam.webproject.entity.user.UserRole;
 import uz.epam.webproject.service.MedicineService;
 import uz.epam.webproject.service.UserService;
 import uz.epam.webproject.service.exception.ServiceException;
@@ -21,27 +22,30 @@ import java.util.Optional;
 
 public class DeleteUserCommand implements Command {
     private static final Logger logger = LogManager.getLogger();
-    private static final String USER_ID = "userId";
 
     @Override
     public Router execute(HttpServletRequest request) throws CommandException {
+        HttpSession session = request.getSession();
         UserService userService = UserServiceImpl.getInstance();
-        long userId = Long.parseLong(request.getParameter(USER_ID));
+        long userId = Long.parseLong(request.getParameter(ParameterName.USER_ID));
         try {
             Optional<User> optionalUser = userService.findById(userId);
             String userToString = "";
             if (optionalUser.isPresent()){
                 userToString =  optionalUser.get().toString();
             }
-            if (userService.delete(userId)){
-                logger.info("User with id " + userId + " successfully deleted");
-                List<User> userList = userService.findAll();
-                request.setAttribute(ParameterName.USERS, userList);
-                request.setAttribute(ParameterName.USER_DELETED, userToString);
-            }else {
-                logger.info("User with id " + userId + " not deleted");
-                request.setAttribute(ParameterName.USER_NOT_DELETED, ParameterName.USER_NOT_DELETED);
+            if (isAdmin(session) || true) {
+                if (userService.delete(userId)) {
+                    logger.info("User with id " + userId + " successfully deleted");
+                    List<User> userList = userService.findAll();
+                    request.setAttribute(ParameterName.USERS, userList);
+                    request.setAttribute(ParameterName.USER_DELETED, userToString);
+                } else {
+                    logger.info("User with id " + userId + " not deleted");
+                    request.setAttribute(ParameterName.USER_NOT_DELETED, ParameterName.USER_NOT_DELETED);
+                }
             }
+            session.setAttribute(ParameterName.CURRENT_PAGE, ParameterName.BOOTSTRAP_USERS_LIST_TABLE);
             return new Router(ParameterName.BOOTSTRAP_USERS_LIST_TABLE);
 
         } catch (ServiceException e) {
@@ -51,8 +55,8 @@ public class DeleteUserCommand implements Command {
     }
 
     @Override
-    public boolean isPharmacist(HttpSession session) {
-        return false;
+    public boolean isAdmin(HttpSession session) {
+        return  session.getAttribute(ParameterName.ROLE).equals(UserRole.ADMIN.toString());
     }
 
     @Override
